Tratamento de excessoes
Como foi exigido tratamento de exceções, mudei a forma como sao tratados os erros em caso de argumentos invalidos.
